ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.35.0
torch==2.6.0
pytesseract==0.3.10
Pillow==11.1.0
scikit-learn==1.3.2
numpy==1.26.2
